+++ klassez/fit.py
@@ -24,7 +24,7 @@ Functions for performing fits.
 s_colors = ['tab:cyan', 'tab:red', 'tab:green', 'tab:purple', 'tab:pink', 'tab:gray', 'tab:brown', 'tab:olive', 'salmon', 'indigo']
 
 
-def histogram(data, nbins=100, density=True, f_lims=None, xlabel=None, x_symm=False, fitG=True, barcolor='tab:blue', fontsize=10, filename=None, ext='png', dpi=600):
+def histogram(data, nbins=100, density=True, f_lims=None, xlabel=None, x_symm=False, fitG=True, barcolor='tab:blue', fontsize=10, filename=None, ext='svg', dpi=600):
     """
     Computes an histogram of ``data`` and tries to fit it with a gaussian lineshape.
     The parameters of the gaussian function are calculated analytically directly from ``data`` using ``scipy.stats.norm``
@@ -396,7 +396,7 @@ def make_signal(t, u, s, k, b, phi, A, SFO1=701.125, o1p=0, N=None):
 
 def plot_fit(S, ppm_scale, regions, t_AQ, SFO1, o1p, show_total=False,
              show_res=False, res_offset=0, show_basl=False, X_label=r'$\delta$ /ppm',
-             labels=None, filename='fit', ext='png', dpi=600, dim=None):
+             labels=None, filename='fit', ext='svg', dpi=600, dim=None):
     """
     Plots either the initial guess or the result of the fit, and saves all the figures.
     The figure `<filename>_full` will show the whole model and the whole spectrum.
@@ -2167,7 +2167,7 @@ class Voigt_Fit:
         """
         gui.edit_vf(self.S, self.ppm_scale, regions, self.t_AQ, self.SFO1, self.o1p, filename=filename, ext=ext)
 
-    def plot(self, what='result', show_total=True, show_res=False, res_offset=0, show_basl=False, labels=None, filename=None, ext='png', dpi=600, dim=None):
+    def plot(self, what='result', show_total=True, show_res=False, res_offset=0, show_basl=False, labels=None, filename=None, ext='svg', dpi=600, dim=None):
         """
         Plots either the initial guess or the result of the fit, and saves all the figures. Calls :func:`fit.plot_fit`.
         The figure `<filename>_full` will show the whole model and the whole spectrum.
@@ -2287,7 +2287,7 @@ class Voigt_Fit:
         total = np.sum(signals, axis=0)
         return signals, total, limits_list, whole_basl
 
-    def res_histogram(self, what='result', nbins=500, density=True, f_lims=None, xlabel='Residuals', x_symm=True, barcolor='tab:green', fontsize=20, filename=None, ext='png', dpi=300):
+    def res_histogram(self, what='result', nbins=500, density=True, f_lims=None, xlabel='Residuals', x_symm=True, barcolor='tab:green', fontsize=20, filename=None, ext='svg', dpi=300):
         """
         Computes the histogram of the residuals and saves it.
         Employs :func:`klassez.fit.histogram` to make the figure.
@@ -2789,7 +2789,7 @@ class Voigt_Fit_2D:
             if len(self.label_list) < len(self.coord):
                 raise ValueError('The number of provided labels is not enough for the peaks.')
 
-    def draw_coord(self, filename=None, labelsize=8, ext='png', dpi=600, **kwargs):
+    def draw_coord(self, filename=None, labelsize=8, ext='svg', dpi=600, **kwargs):
         """
         Makes a figure with the experimental dataset and the peak-picked signals as crosshairs.
 
@@ -3883,7 +3883,7 @@ def make_iguess_dosy(x, labels, data, model, model_args, diff_c_0=1e-10, filenam
     print(f'\n{filename}.idy saved.', c='tab:blue')
 
 
-def plot_fit_P2D(S, ppm_scale, regions, t_AQ, SFO1, o1p, show_total=False, show_res=False, res_offset=0, X_label=r'$\delta$ /ppm', labels=None, filename='fit', ext='png', dpi=600):
+def plot_fit_P2D(S, ppm_scale, regions, t_AQ, SFO1, o1p, show_total=False, show_res=False, res_offset=0, X_label=r'$\delta$ /ppm', labels=None, filename='fit', ext='svg', dpi=600):
     """
     Plots either the initial guess or the result of the fit, and saves all the figures.
     A new folder named <filename>_fit will be created.
@@ -4470,7 +4470,7 @@ class Voigt_Fit_P2D:
         # Store
         self.result = fit.read_vf_P2D(f'{filename}.fvf')
 
-    def plot(self, what='result', show_total=True, show_res=False, res_offset=0, labels=None, filename=None, ext='png', dpi=600):
+    def plot(self, what='result', show_total=True, show_res=False, res_offset=0, labels=None, filename=None, ext='svg', dpi=600):
         """
         Plots either the initial guess or the result of the fit, and saves all the figures. Calls fit.plot_fit_P2D.
         The figures <filename>_full will show the whole model and the whole spectrum.
@@ -4573,7 +4573,7 @@ class Voigt_Fit_P2D:
         total = np.sum(signals, axis=1)  # sum the peaks
         return signals, total, limits_list
 
-    def res_histogram(self, what='result', nbins=500, density=True, f_lims=None, xlabel='Residuals', x_symm=True, barcolor='tab:green', fontsize=20, filename=None, ext='png', dpi=300):
+    def res_histogram(self, what='result', nbins=500, density=True, f_lims=None, xlabel='Residuals', x_symm=True, barcolor='tab:green', fontsize=20, filename=None, ext='svg', dpi=300):
         """
         Computes the histogram of the residuals and saves it in the same folder of the fit figures.
         Employs fit.histogram to make the figure.
@@ -4964,7 +4964,7 @@ def fit_dosy_multi(x, y, iguess, model, model_args, d_bds=3, f_bds=[0, 3], vary_
     return dic_result
 
 
-def plot_fit_dosy(x, label, y, total, yc, region, show_total=True, show_res=False, res_offset=0, filename=None, ext='png', dpi=600, dim=None):
+def plot_fit_dosy(x, label, y, total, yc, region, show_total=True, show_res=False, res_offset=0, filename=None, ext='svg', dpi=600, dim=None):
     """
     Make a plot of a DOSY fit.
 
@@ -5057,7 +5057,7 @@ def plot_fit_dosy(x, label, y, total, yc, region, show_total=True, show_res=Fals
     plt.close()
 
 
-def plot_fit_dosy_multi(x, yy, totals, components, region, bigdeltas=None, colors=None, filename=None, ext='png', dpi=300, dim=None):
+def plot_fit_dosy_multi(x, yy, totals, components, region, bigdeltas=None, colors=None, filename=None, ext='svg', dpi=300, dim=None):
     """
     Makes a cumulative plot of a DOSY fit performed with the :class:`klassez.fit.DosyFit_pp3D` class.
 
@@ -5414,10 +5414,10 @@ class DosyFit:
         self.result = regions
         print(f'{path_x} loaded as fit result file.\n', c='tab:blue')
 
-    def plot(self, what='result', show_res=False, res_offset=0, filename=None, ext='png', dpi=600, dim=None):
+    def plot(self, what='result', show_res=False, res_offset=0, filename=None, ext='svg', dpi=600, dim=None):
         """
         Plots either the initial guess or the result of the fit, and saves all the figures. Calls :func:`fit.plot_fit_dosy`.
-        The figures will be saved in the directory `Figures_<filename>/<what>/<label>.png`.
+        The figures will be saved in the directory `Figures_<filename>/<what>/<label>.svg`.
 
         Parameters
         ----------
@@ -6151,10 +6151,10 @@ class DosyFit_pp3D(fit.DosyFit):
 
         self.merge_planes('result')
 
-    def plot(self, what='result', only_all=False, show_res=False, res_offset=0, figdir=None, filename=None, ext='png', dpi=100, dim=None):
+    def plot(self, what='result', only_all=False, show_res=False, res_offset=0, figdir=None, filename=None, ext='svg', dpi=100, dim=None):
         """
         Plots either the initial guess or the result of the fit, and saves all the figures. Calls :func:`fit.plot_fit_dosy`.
-        The figures will be saved in the directory `<figdir>/<what>/<label>.png`.
+        The figures will be saved in the directory `<figdir>/<what>/<label>.<ext>`.
 
         Parameters
         ----------